+++ Assets/Scenes/Playground.unity
@@ -1008,6 +1008,7 @@ GameObject:
   m_Component:
   - component: {fileID: 1682951153}
   - component: {fileID: 1682951152}
+  - component: {fileID: 1682951156}
   - component: {fileID: 1682951155}
   - component: {fileID: 1682951151}
   - component: {fileID: 1682951154}
@@ -1136,6 +1137,23 @@ MonoBehaviour:
   minSize: 2.8
   zoomSpeed: 1
   zoomDamping: 2
+--- !u!114 &1682951156
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1682951150}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 29f4b1f6be0c6bc44ad309c5547dd639, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  target: {fileID: 981443859}
+  input: {fileID: 981443860}
+  offset: {x: 0, y: 13, z: -5.5}
+  aheadSpeed: 3
+  followDamping: 0.5
 --- !u!1001 &1720029418
 PrefabInstance:
   m_ObjectHideFlags: 0
