+++ Assets/Scripts/Camera/FollowingCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29f4b1f6be0c6bc44ad309c5547dd639
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
